:broom: Extracting drivers and simplification of Vec2 (#208)
Extract drivers out into their own file instead of `ui.tsx`.

Merge some duplicated code.

Also, use a new type `Path` instead of `Vec2[] & { stroke?: string, groupId?: string })`.  Make `stroke` and `groupId` mandatory for this type.  Any calculations done without them should be on `Vec2[]` to make it clear the information is lost.
